feat: creates a hook for the file download and preview
