+++ app/components/files/document_file.tsx
@@ -1,14 +1,19 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-import React, {forwardRef, useImperativeHandle, useRef, useState} from 'react';
+import React, {forwardRef, useImperativeHandle, useRef} from 'react';
 import {StyleSheet, TouchableOpacity, View} from 'react-native';
 
 import Document, {type DocumentRef} from '@components/document';
 import ProgressBar from '@components/progress_bar';
 import {useTheme} from '@context/theme';
+import {useDownloadFileAndPreview} from '@hooks/files';
 
 import FileIcon from './file_icon';
+
+export type DocumentFileRef = {
+    handlePreviewPress: () => void;
+}
 
 type DocumentFileProps = {
     backgroundColor?: string;
@@ -29,8 +34,8 @@ const styles = StyleSheet.create({
 
 const DocumentFile = forwardRef<DocumentRef, DocumentFileProps>(({backgroundColor, canDownloadFiles, disabled = false, file}: DocumentFileProps, ref) => {
     const theme = useTheme();
-    const [progress, setProgress] = useState(0);
     const document = useRef<DocumentRef>(null);
+    const {downloading, progress, toggleDownloadAndPreview} = useDownloadFileAndPreview();
 
     const handlePreviewPress = async () => {
         document.current?.handlePreviewPress();
@@ -48,7 +53,7 @@ const DocumentFile = forwardRef<DocumentRef, DocumentFileProps>(({backgroundColo
     );
 
     let fileAttachmentComponent = icon;
-    if (progress) {
+    if (downloading) {
         fileAttachmentComponent = (
             <>
                 {icon}
@@ -66,7 +71,7 @@ const DocumentFile = forwardRef<DocumentRef, DocumentFileProps>(({backgroundColo
         <Document
             canDownloadFiles={canDownloadFiles}
             file={file}
-            onProgress={setProgress}
+            downloadAndPreviewFile={toggleDownloadAndPreview}
             ref={document}
         >
             <TouchableOpacity
