+++ app/hooks/files.ts
@@ -1,14 +1,24 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-import {useEffect, useMemo, useState} from 'react';
+import {deleteAsync} from 'expo-file-system';
+import {useEffect, useMemo, useRef, useState} from 'react';
+import {useIntl} from 'react-intl';
+import {Platform, StatusBar, type StatusBarStyle} from 'react-native';
+import FileViewer from 'react-native-file-viewer';
+import tinyColor from 'tinycolor2';
 
 import {getLocalFileInfo} from '@actions/local/file';
-import {buildFilePreviewUrl, buildFileUrl} from '@actions/remote/file';
+import {buildFilePreviewUrl, buildFileUrl, downloadFile} from '@actions/remote/file';
 import {useServerUrl} from '@context/server';
-import {isAudio, isGif, isImage, isVideo} from '@utils/file';
+import {useTheme} from '@context/theme';
+import {alertDownloadFailed, alertFailedToOpenDocument} from '@utils/document';
+import {getFullErrorMessage, isErrorWithMessage} from '@utils/errors';
+import {fileExists, getLocalFilePathFromFile, isAudio, isGif, isImage, isVideo} from '@utils/file';
 import {getImageSize} from '@utils/gallery';
+import {logDebug} from '@utils/log';
 
+import type {ClientResponse, ProgressPromise} from '@mattermost/react-native-network-client';
 import type ChannelBookmarkModel from '@typings/database/models/servers/channel_bookmark';
 
 const getFileInfo = async (serverUrl: string, bookmarks: ChannelBookmarkModel[], publicLinkEnabled: boolean, cb: (files: FileInfo[]) => void) => {
@@ -87,4 +97,136 @@ export const useChannelBookmarkFiles = (bookmarks: ChannelBookmarkModel[], publi
     }, [serverUrl, bookmarks, publicLinkEnabled]);
 
     return files;
+};
+
+export const useDownloadFileAndPreview = () => {
+    const serverUrl = useServerUrl();
+    const intl = useIntl();
+    const theme = useTheme();
+    const downloadTask = useRef<ProgressPromise<ClientResponse>>();
+
+    const [progress, setProgress] = useState<number>(0);
+
+    const [preview, setPreview] = useState(false);
+    const [downloading, setDownloading] = useState(false);
+    const [didCancel, setDidCancel] = useState(false);
+
+    const openDocument = async (file: FileInfo) => {
+        if (!didCancel && !preview) {
+            let path = decodeURIComponent(file.localPath || '');
+            let exists = false;
+            if (path) {
+                exists = await fileExists(path);
+            }
+
+            if (!exists) {
+                path = getLocalFilePathFromFile(serverUrl, file);
+            }
+
+            setPreview(true);
+            setStatusBarColor('dark-content');
+            FileViewer.open(path!.replace('file://', ''), {
+                displayName: decodeURIComponent(file.name),
+                onDismiss: onDonePreviewingFile,
+                showOpenWithDialog: true,
+                showAppsSuggestions: true,
+            }).then(() => {
+                setDownloading(false);
+                setProgress(0);
+            }).catch(() => {
+                alertFailedToOpenDocument(file, intl);
+                onDonePreviewingFile();
+
+                if (path) {
+                    deleteAsync(path, {idempotent: true});
+                }
+            });
+        }
+    };
+
+    const downloadAndPreviewFile = async (file: FileInfo) => {
+        setDidCancel(false);
+        let path;
+        let exists = false;
+
+        try {
+            path = decodeURIComponent(file.localPath || '');
+            if (path) {
+                exists = await fileExists(path);
+            }
+
+            if (!exists) {
+                path = getLocalFilePathFromFile(serverUrl, file);
+                exists = await fileExists(path);
+            }
+
+            if (exists) {
+                openDocument(file);
+            } else {
+                setDownloading(true);
+                downloadTask.current = downloadFile(serverUrl, file.id!, path!);
+                downloadTask.current?.progress?.(setProgress);
+
+                await downloadTask.current;
+                setProgress(1);
+                openDocument(file);
+            }
+        } catch (error) {
+            if (path) {
+                deleteAsync(path, {idempotent: true});
+            }
+            setDownloading(false);
+            setProgress(0);
+
+            if (!isErrorWithMessage(error) || error.message !== 'cancelled') {
+                logDebug('error on downloadAndPreviewFile', getFullErrorMessage(error));
+                alertDownloadFailed(intl);
+            }
+        }
+    };
+
+    const setStatusBarColor = (style: StatusBarStyle = 'light-content') => {
+        if (Platform.OS === 'ios') {
+            if (style) {
+                StatusBar.setBarStyle(style, true);
+            } else {
+                const headerColor = tinyColor(theme.sidebarHeaderBg);
+                let barStyle: StatusBarStyle = 'light-content';
+                if (headerColor.isLight() && Platform.OS === 'ios') {
+                    barStyle = 'dark-content';
+                }
+                StatusBar.setBarStyle(barStyle, true);
+            }
+        }
+    };
+
+    const toggleDownloadAndPreview = (file: FileInfo) => {
+        if (downloading) {
+            setProgress(0);
+            cancelDownload();
+            setDownloading(false);
+        } else {
+            downloadAndPreviewFile(file);
+        }
+    };
+
+    const cancelDownload = () => {
+        setDidCancel(true);
+        if (downloadTask.current?.cancel) {
+            downloadTask.current.cancel();
+        }
+    };
+
+    const onDonePreviewingFile = () => {
+        setProgress(0);
+        setDownloading(false);
+        setPreview(false);
+        setStatusBarColor();
+    };
+
+    return {
+        downloading,
+        progress,
+        toggleDownloadAndPreview,
+    };
 };
